Remove .vs folder, move OSCUtils to new class
